Split devices module into devices and initialize modules
initialize contains solely the initialization of devices and testing
for that, which is a good 600 lines.
